+++ src/it/scala/co/topl/brambl/cli/ComplexTransactionWithFileTest.scala
@@ -50,7 +50,7 @@ class ComplexTransactionWithFileTest
             .iterateUntil(_ == ExitCode.Success),
           60.seconds
         )
-        ALICE_TO_ADDRESS <- walletController(ALICE_WALLET).currentaddress()
+        ALICE_TO_ADDRESS <- walletController(ALICE_WALLET).currentaddress("self", "default", None)
         genesisAddress <- walletController(ALICE_WALLET)
           .currentaddress("noparty", "genesis", Some(1))
         utxos <- genusQueryAlgebra
@@ -84,7 +84,7 @@ class ComplexTransactionWithFileTest
           genesisAmount - BASE_AMOUNT,
           BASE_FEE,
           genesisAddress.get,
-          ALICE_TO_ADDRESS.toOption.get
+          ALICE_TO_ADDRESS.get
         )
         _ <- assertIO(
           createComplexTransactionToAddress(