+++ src/main/scala/co/topl/brambl/cli/controllers/GenusQueryController.scala
@@ -7,13 +7,15 @@ import co.topl.brambl.codecs.AddressCodecs
 import co.topl.brambl.dataApi.GenusQueryAlgebra
 import topl.brambl.dataApi.WalletStateAlgebra
 import topl.brambl.cli.TokenType
+import cats.effect.kernel.Sync
 
-class GenusQueryController[F[_]: Monad](
+class GenusQueryController[F[_]: Sync](
     walletStateAlgebra: WalletStateAlgebra[F],
     genusQueryAlgebra: GenusQueryAlgebra[F]
 ) {
 
   def queryUtxoFromParams(
+      someFromAddress: Option[String],
       fromParty: String,
       fromContract: String,
       someFromState: Option[Int],
@@ -21,13 +23,17 @@ class GenusQueryController[F[_]: Monad](
   ): F[Either[String, String]] = {
 
     import cats.implicits._
-    walletStateAlgebra
-      .getAddress(fromParty, fromContract, someFromState)
+    someFromAddress
+      .map(x => Sync[F].point(Some(x)))
+      .getOrElse(
+        walletStateAlgebra
+          .getAddress(fromParty, fromContract, someFromState)
+      )
       .flatMap {
         case Some(address) =>
           genusQueryAlgebra
             .queryUtxo(AddressCodecs.decodeAddress(address).toOption.get)
-            .map(_.filter{x => 
+            .map(_.filter { x =>
               import monocle.macros.syntax.lens._
               val lens = x.focus(_.transactionOutput.value.value)
               if (tokenType == TokenType.lvl)
@@ -49,6 +55,13 @@ class GenusQueryController[F[_]: Monad](
                 Right((txos.map { txo =>
                   BlockDisplayOps.display(txo)
                 }).mkString)
+            }
+            .attempt
+            .map {
+              _ match {
+                case Left(_)     => Left("Problem contacting the network.")
+                case Right(txos) => txos
+              }
             }
         case None => Monad[F].pure(Left("Address not found"))
       }