+++ src/main/scala/co/topl/brambl/cli/impl/SimpleMintingAlgebra.scala
@@ -10,6 +10,7 @@ import co.topl.brambl.models.box.AssetMintingStatement
 import co.topl.brambl.wallet.WalletApi
 import com.google.protobuf.ByteString
 import io.circe.Json
+import co.topl.genus.services.Txo
 
 trait SimpleMintingAlgebra[F[_]] {
   def createSimpleGroupMintingTransactionFromParams(
@@ -138,7 +139,18 @@ object SimpleMintingAlgebra {
       fromAddress <- transactionBuilderApi.lockAddress(
         predicateFundsToUnlock
       )
-      response <- utxoAlgebra.queryUtxo(fromAddress)
+      response <- utxoAlgebra
+        .queryUtxo(fromAddress)
+        .attempt
+        .flatMap {
+          _ match {
+            case Left(_) =>
+              Sync[F].raiseError(
+                CreateTxError("Problem contacting network")
+              ): F[Seq[Txo]]
+            case Right(txos) => Sync[F].pure(txos: Seq[Txo])
+          }
+        }
       lvlTxos = response.filter(
         _.transactionOutput.value.value.isLvl
       )
@@ -185,7 +197,18 @@ object SimpleMintingAlgebra {
       fromAddress <- transactionBuilderApi.lockAddress(
         predicateFundsToUnlock
       )
-      response <- utxoAlgebra.queryUtxo(fromAddress)
+      response <- utxoAlgebra
+        .queryUtxo(fromAddress)
+        .attempt
+        .flatMap {
+          _ match {
+            case Left(_) =>
+              Sync[F].raiseError(
+                CreateTxError("Problem contacting network")
+              ): F[Seq[Txo]]
+            case Right(txos) => Sync[F].pure(txos: Seq[Txo])
+          }
+        }
       lvlTxos = response.filter(
         _.transactionOutput.value.value.isLvl
       )
@@ -233,7 +256,18 @@ object SimpleMintingAlgebra {
       fromAddress <- transactionBuilderApi.lockAddress(
         predicateFundsToUnlock
       )
-      response <- utxoAlgebra.queryUtxo(fromAddress)
+      response <- utxoAlgebra
+        .queryUtxo(fromAddress)
+        .attempt
+        .flatMap {
+          _ match {
+            case Left(_) =>
+              Sync[F].raiseError(
+                CreateTxError("Problem contacting network")
+              ): F[Seq[Txo]]
+            case Right(txos) => Sync[F].pure(txos: Seq[Txo])
+          }
+        }
       lvlTxos = response.filter(
         _.transactionOutput.value.value.isLvl
       )