+++ src/main/scala/co/topl/brambl/cli/impl/SimpleTransactionAlgebra.scala
@@ -203,7 +203,18 @@ object SimpleTransactionAlgebra {
           fromAddress <- transactionBuilderApi.lockAddress(
             predicateFundsToUnlock.get
           )
-          response <- utxoAlgebra.queryUtxo(fromAddress)
+          response <- utxoAlgebra
+            .queryUtxo(fromAddress)
+            .attempt
+            .flatMap {
+              _ match {
+                case Left(_) =>
+                  Sync[F].raiseError(
+                    CreateTxError("Problem contacting network")
+                  ): F[Seq[Txo]]
+                case Right(txos) => Sync[F].pure(txos: Seq[Txo])
+              }
+            }
           txos = response
             .filter(x =>
               !x.transactionOutput.value.value.isTopl &&
@@ -261,7 +272,7 @@ object SimpleTransactionAlgebra {
           e match {
             case Right(_)                               => ().asRight
             case Left(e: SimpleTransactionAlgebraError) => e.asLeft
-            case Left(e) => 
+            case Left(e) =>
               e.printStackTrace()
               UnexpectedError(e.getMessage()).asLeft
           }