+++ src/main/scala/co/topl/brambl/cli/modules/ChannelResourceModule.scala
@@ -5,14 +5,23 @@ import cats.effect.kernel.Sync
 import io.grpc.ManagedChannelBuilder
 
 trait ChannelResourceModule {
-  def channelResource[F[_]: Sync](address: String, port: Int) = {
+  def channelResource[F[_]: Sync](
+      address: String,
+      port: Int,
+      secureConnection: Boolean
+  ) = {
     Resource
       .make {
         Sync[F].delay(
-          ManagedChannelBuilder
-            .forAddress(address, port)
-            .usePlaintext()
-            .build
+          if (secureConnection)
+            ManagedChannelBuilder
+              .forAddress(address, port)
+              .build
+          else
+            ManagedChannelBuilder
+              .forAddress(address, port)
+              .usePlaintext()
+              .build
         )
       }(channel => Sync[F].delay(channel.shutdown()))
   }