+++ src/main/scala/co/topl/brambl/cli/modules/TxModeModule.scala
@@ -23,7 +23,8 @@ trait TxModeModule extends TxParserAlgebraModule with TransactionAlgebraModule {
           transactionOps(
             validateParams.walletFile,
             validateParams.host,
-            validateParams.bifrostPort
+            validateParams.bifrostPort,
+            validateParams.secureConnection
           )
         ).broadcastSimpleTransactionFromParams(validateParams.someInputFile.get)
       case BramblCliSubCmd.prove =>
@@ -35,7 +36,8 @@ trait TxModeModule extends TxParserAlgebraModule with TransactionAlgebraModule {
           transactionOps(
             validateParams.walletFile,
             validateParams.host,
-            validateParams.bifrostPort
+            validateParams.bifrostPort,
+            validateParams.secureConnection
           )
         ).proveSimpleTransactionFromParams(
           validateParams.someInputFile.get,
@@ -52,7 +54,8 @@ trait TxModeModule extends TxParserAlgebraModule with TransactionAlgebraModule {
           transactionOps(
             validateParams.walletFile,
             validateParams.host,
-            validateParams.bifrostPort
+            validateParams.bifrostPort,
+            validateParams.secureConnection
           )
         ).inspectTransaction(validateParams.someInputFile.get)
       case BramblCliSubCmd.create =>
@@ -64,7 +67,8 @@ trait TxModeModule extends TxParserAlgebraModule with TransactionAlgebraModule {
           transactionOps(
             validateParams.walletFile,
             validateParams.host,
-            validateParams.bifrostPort
+            validateParams.bifrostPort,
+            validateParams.secureConnection
           )
         )
           .createComplexTransaction(